fix(tools): self-heal TS6305 stale-artifact false red in check:local
